+++ lib/instream/query/write.ex
@@ -5,13 +5,17 @@ defmodule Instream.Query.Write do
 
   use Instream.Query
 
+  alias Instream.Query.Headers
+  alias Instream.Query.URL
+
   def execute(%Query{ payload: payload }, opts, conn) do
-    url =
+    headers = conn |> Headers.assemble()
+    url     =
          conn
       |> URL.write()
       |> URL.append_database(opts[:database])
 
-    { :ok, _, _, client } = :hackney.post(url, [], payload)
+    { :ok, _, _, client } = :hackney.post(url, headers, payload)
     { :ok, response }     = :hackney.body(client)
 
     response |> maybe_parse(opts)