+++ test/helpers/query_auth_connection.exs
@@ -1,0 +1,3 @@
+defmodule Instream.TestHelpers.QueryAuthConnection do
+  use Instream.Connection, otp_app: :instream_test
+end