+++ test/instream/query/headers_test.exs
@@ -1,0 +1,5 @@
+defmodule Instream.Query.HeadersTest do
+  use ExUnit.Case, async: true
+
+  doctest Instream.Query.Headers, import: true
+end